requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
cx_Oracle==8.3.0
flask-wtf==1.2.2
Flask-Login==0.6.3